+++ PerformWorld/src/main/resources/templates/admin/AdminUserList.html
@@ -48,11 +48,6 @@
       <!-- User 조회 섹션 -->
       <div class="col-md-12">
         <h2>User 조회</h2>
-        <div class="row">
-          <div class="col-md-12 text-end">
-            <button class="btn btn-primary add-user-btn">Add User</button>
-          </div>
-        </div>
         <!-- User 조회 테이블 -->
         <div class="row">
           <div class="col-md-12">
@@ -69,7 +64,7 @@
               </tr>
               </thead>
               <tbody>
-              <tr th:each="user : ${users}">
+              <tr th:each="user : ${users}" onclick="openEditUserModal(${user.userId})">
                 <td th:text="${user.userId}"></td>
                 <td th:text="${user.name}"></td>
                 <td th:text="${user.email}"></td>
@@ -113,59 +108,6 @@
               <input type="number" class="form-control" id="discountRate" required>
             </div>
             <button type="submit" class="btn btn-primary">Tier 추가</button>
-          </form>
-        </div>
-      </div>
-    </div>
-  </div>
-
-  <!-- 사용자 추가 모달 -->
-  <div id="addUserModal" class="modal" style="display: none;">
-    <div class="modal-dialog">
-      <div class="modal-content">
-        <div class="modal-header" style="text-align: right;">
-          <h4 class="modal-title">사용자 추가</h4>
-          <button type="button" class="close" onclick="closeAddUserModal()">&times;</button>
-        </div>
-        <div class="modal-body" style="max-height: 400px; overflow-y: auto;">
-          <form id="addUserForm">
-            <!-- 기존 폼 필드들 -->
-            <div class="form-group">
-              <label for="userId">User ID</label>
-              <input type="text" class="form-control" id="userId" required>
-            </div>
-            <div class="form-group">
-              <label for="name">이름</label>
-              <input type="text" class="form-control" id="name" required>
-            </div>
-            <div class="form-group">
-              <label for="email">이메일</label>
-              <input type="email" class="form-control" id="email" required>
-            </div>
-            <div class="form-group">
-              <label for="password">비밀번호</label>
-              <input type="password" class="form-control" id="password" required>
-            </div>
-            <div class="form-group">
-              <label for="phoneNumber">전화번호</label>
-              <input type="text" class="form-control" id="phoneNumber" required>
-            </div>
-            <div class="form-group">
-              <label for="address1">주소 1</label>
-              <input type="text" class="form-control" id="address1" required>
-            </div>
-            <div class="form-group">
-              <label for="address2">주소 2</label>
-              <input type="text" class="form-control" id="address2" required>
-            </div>
-            <div class="form-group">
-              <label for="postcode">우편번호</label>
-              <input type="text" class="form-control" id="postcode" required>
-            </div>
-            <div class="form-group">
-              <input type="hidden" id="tierId" name="tierId" value="1">
-            </div>
-            <button type="submit" class="btn btn-primary">사용자 추가</button>
           </form>
         </div>
       </div>
